+++ lib/bugsnag/configuration.rb
@@ -199,9 +199,9 @@ module Bugsnag
 
     DEFAULT_NOTIFY_ENDPOINT = "https://notify.bugsnag.com"
     DEFAULT_SESSION_ENDPOINT = "https://sessions.bugsnag.com"
-    HUB_NOTIFY_ENDPOINT = "https://notify.insighthub.smartbear.com"
-    HUB_SESSION_ENDPOINT = "https://sessions.insighthub.smartbear.com"
-    HUB_PREFIX = "00000"
+    SECONDARY_NOTIFY_ENDPOINT = "https://notify.bugsnag.smartbear.com"
+    SECONDARY_SESSION_ENDPOINT = "https://sessions.bugsnag.smartbear.com"
+    SECONDARY_PREFIX = "00000"
 
     DEFAULT_META_DATA_FILTERS = [
       /authorization/i,
@@ -548,8 +548,8 @@ module Bugsnag
     def set_default_endpoints
       return unless @endpoints.notify.nil? && @endpoints.sessions.nil?
 
-      self.endpoints = if hub_api_key?
-                         EndpointConfiguration.new(HUB_NOTIFY_ENDPOINT, HUB_SESSION_ENDPOINT)
+      self.endpoints = if secondary_api_key?
+                         EndpointConfiguration.new(SECONDARY_NOTIFY_ENDPOINT, SECONDARY_SESSION_ENDPOINT)
                        else
                          EndpointConfiguration.new(DEFAULT_NOTIFY_ENDPOINT, DEFAULT_SESSION_ENDPOINT)
                        end
@@ -769,8 +769,8 @@ module Bugsnag
       ENV["DYNO"] || Socket.gethostname;
     end
 
-    def hub_api_key?
-      @api_key && @api_key.start_with?(HUB_PREFIX)
+    def secondary_api_key?
+      @api_key && @api_key.start_with?(SECONDARY_PREFIX)
     end
   end
 end